requirements.txt
Django==1.8.4
python-social-auth
django-mptt
